Added get-elering-lowest-price.yml
Added a get-elering-lowest-price.yml file in the DSL/Ruuter.public/elering/GET path. This service enables the user to request today's lowest energy price for Estonia. The service first uses worldtimeapi.org api to get today's date and the timzone offset for Tallinn, which are then used for the request against elering energy price api. Then a sorting algorithm is used to get the cheapest energy price today and the time at that point, which is then converted to Tallinn local time.